# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Pecuaria API",
        "description": "Backend do sistema de gestao pecuaria (multi-fazenda) \u00e2\u20ac\u201d painel + motor de gatilhos.",
        "version": "0.1.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "description": "Login via form (username = email). Retorna JWT.",
                "operationId": "login_auth_login_post",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_login_auth_login_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Me",
                "operationId": "me_auth_me_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioOut"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/fazendas": {
            "get": {
                "tags": [
                    "fazendas"
                ],
                "summary": "Listar Fazendas",
                "operationId": "listar_fazendas_fazendas_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/FazendaOut"
                                    },
                                    "type": "array",
                                    "title": "Response Listar Fazendas Fazendas Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/fazendas/{fazenda_id}/parametros": {
            "get": {
                "tags": [
                    "fazendas"
                ],
                "summary": "Parametros Da Fazenda",
                "operationId": "parametros_da_fazenda_fazendas__fazenda_id__parametros_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "fazenda_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Fazenda Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ParametroOut"
                                    },
                                    "title": "Response Parametros Da Fazenda Fazendas  Fazenda Id  Parametros Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/fazendas/{fazenda_id}/parametros/{chave}": {
            "put": {
                "tags": [
                    "fazendas"
                ],
                "summary": "Atualizar Parametro",
                "operationId": "atualizar_parametro_fazendas__fazenda_id__parametros__chave__put",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "fazenda_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Fazenda Id"
                        }
                    },
                    {
                        "name": "chave",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Chave"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ParametroUpdateIn"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParametroOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/dashboard": {
            "get": {
                "tags": [
                    "dashboard"
                ],
                "summary": "Dashboard Consolidado",
                "description": "Painel consolidado do grupo + cartoes por fazenda (no escopo do usuario).",
                "operationId": "dashboard_consolidado_dashboard_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DashboardOut"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/dashboard/fazenda/{fazenda_id}": {
            "get": {
                "tags": [
                    "dashboard"
                ],
                "summary": "Painel Fazenda",
                "operationId": "painel_fazenda_dashboard_fazenda__fazenda_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "fazenda_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Fazenda Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FazendaPainelOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/alertas": {
            "get": {
                "tags": [
                    "alertas"
                ],
                "summary": "Listar Alertas",
                "description": "Alertas de todas as fazendas no escopo do usuario, mais severos primeiro.",
                "operationId": "listar_alertas_alertas_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "apenas_abertos",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true,
                            "title": "Apenas Abertos"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AlertaOut"
                                    },
                                    "title": "Response Listar Alertas Alertas Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/alertas/reavaliar/{fazenda_id}": {
            "post": {
                "tags": [
                    "alertas"
                ],
                "summary": "Reavaliar Fazenda",
                "description": "Roda o motor de gatilhos para a fazenda e retorna os alertas abertos resultantes.",
                "operationId": "reavaliar_fazenda_alertas_reavaliar__fazenda_id__post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "fazenda_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Fazenda Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AlertaOut"
                                    },
                                    "title": "Response Reavaliar Fazenda Alertas Reavaliar  Fazenda Id  Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_login_auth_login_post": {
                "properties": {
                    "grant_type": {
                        "anyOf": [
                            {
                                "type": "string",
                                "pattern": "password"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Grant Type"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    },
                    "scope": {
                        "type": "string",
                        "title": "Scope",
                        "default": ""
                    },
                    "client_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Client Id"
                    },
                    "client_secret": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Client Secret"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "password"
                ],
                "title": "Body_login_auth_login_post"
            },
            "DashboardOut": {
                "properties": {
                    "organizacao": {
                        "type": "string",
                        "title": "Organizacao"
                    },
                    "resumo": {
                        "$ref": "#/components/schemas/ResumoDashboard"
                    },
                    "fazendas": {
                        "items": {
                            "$ref": "#/components/schemas/FazendaPainelOut"
                        },
                        "type": "array",
                        "title": "Fazendas"
                    }
                },
                "type": "object",
                "required": [
                    "organizacao",
                    "resumo",
                    "fazendas"
                ],
                "title": "DashboardOut"
            },
            "FazendaOut": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "nome": {
                        "type": "string",
                        "title": "Nome"
                    },
                    "municipio": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Municipio"
                    },
                    "uf": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Uf"
                    },
                    "ativo": {
                        "type": "boolean",
                        "title": "Ativo"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "nome",
                    "municipio",
                    "uf",
                    "ativo"
                ],
                "title": "FazendaOut"
            },
            "FazendaPainelOut": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "nome": {
                        "type": "string",
                        "title": "Nome"
                    },
                    "municipio": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Municipio"
                    },
                    "uf": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Uf"
                    },
                    "alertas_abertos": {
                        "type": "integer",
                        "title": "Alertas Abertos"
                    },
                    "por_severidade": {
                        "additionalProperties": {
                            "type": "integer"
                        },
                        "type": "object",
                        "title": "Por Severidade"
                    },
                    "indicadores": {
                        "items": {
                            "$ref": "#/components/schemas/IndicadorPainelOut"
                        },
                        "type": "array",
                        "title": "Indicadores"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "nome",
                    "municipio",
                    "uf",
                    "alertas_abertos",
                    "por_severidade",
                    "indicadores"
                ],
                "title": "FazendaPainelOut"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ParametroOut": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "grupo": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Grupo"
                    },
                    "chave": {
                        "type": "string",
                        "title": "Chave"
                    },
                    "rotulo": {
                        "type": "string",
                        "title": "Rotulo"
                    },
                    "valor": {
                        "type": "number",
                        "title": "Valor"
                    },
                    "unidade": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Unidade"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "grupo",
                    "chave",
                    "rotulo",
                    "valor",
                    "unidade"
                ],
                "title": "ParametroOut"
            },
            "ParametroUpdateIn": {
                "properties": {
                    "valor": {
                        "type": "number",
                        "title": "Valor"
                    }
                },
                "type": "object",
                "required": [
                    "valor"
                ],
                "title": "ParametroUpdateIn"
            },
...
            "TokenOut": {
                "properties": {
                    "access_token": {
                        "type": "string",
                        "title": "Access Token"
                    },
                    "token_type": {
                        "type": "string",
                        "title": "Token Type",
                        "default": "bearer"
                    }
                },
                "type": "object",
                "required": [
                    "access_token"
                ],
                "title": "TokenOut"
            },
            "UsuarioOut": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "nome": {
                        "type": "string",
                        "title": "Nome"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "papel": {
                        "type": "string",
                        "title": "Papel"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "nome",
                    "email",
                    "papel"
                ],
                "title": "UsuarioOut"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/auth/login"
                    }
                }
            }
        }
    }
}